Created a Keepalive Azure Function
Based on a 4 minute timer this function will run and call my website. I
created one of these already in Python, but it requires that I have my
computer running all the time, and sometimes I do maintenance etc... so
this is a nice solution. There is a bit of overkill in this Azure
Function around calling the web page, it was old code I had lying
around, so I cut-n-pasted it, so it was fast for me. I could clean that
up and I could add more functionality in the future, but for now this is
good enough.