Refactored the GetUrl method off of the model objects
It didn’t make sense to have GetUrl on IPopulateRecord, plus I want a
generic way of getting the urls strings from configuration based on the
type of object. So I moved the method up into PaperRepository.

This is a MUCH cleaner way to go. Now GetRecords takes a Url. I don’t
have to new up a type T to be able to get the URL … so removing that
hack.